# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  git_neolidy_top_neo_flowx_pkg_response.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_interfaces_http_handler.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  internal_interfaces_http_handler.registerRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      tenant_id:
        maxLength: 26
        minLength: 1
        type: string
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - password
    - tenant_id
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: neolidy@outlook.com
    name: jiangfire
    url: https://github.com/jiangfire/flowx
  description: 企业级智能工具治理与流程编排平台 API 文档
  license:
    name: MIT
    url: https://github.com/jiangfire/flowx/blob/main/LICENSE
  termsOfService: https://github.com/jiangfire/flowx
  title: FlowX API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录获取令牌
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      summary: 用户登录
      tags:
      - 认证
  /auth/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 注册请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"